upload started

+++ R/docker.R
@@ -2,6 +2,7 @@ docker_cmd <- function(args, std_out = TRUE, std_err = TRUE) {
   outsider.base::is_docker_available()
   cmd_args <- crayon::silver(paste0('docker ', paste(args, collapse = ' ')))
   cat_line(crayon::bold('Command:\n'), cmd_args)
+  cat_line(crayon::silver(cli::rule(line = '.')))
   # run through callr
   callr_args <- list(args, std_out, std_err)
   res <- callr::r(func = function(args, std_out, std_err) {
@@ -16,12 +17,20 @@ docker_build <- function(img, tag, url_or_path, verbose) {
   docker_cmd(args = args, std_out = verbose, std_err = verbose)
 }
 
-docker_push <- function(img, tag, verbose) {
+docker_push <- function(username, img, tag, verbose) {
+  if (!docker_login(username = username)) {
+    stop(paste0('Unable to login to DockerHub with username ', char(username)),
+         call. = FALSE)
+  }
   args <- c('push', paste0(img, ':', tag))
   docker_cmd(args = args, std_out = verbose, std_err = verbose)
 }
 
 docker_login <- function(username) {
+  res <- sys::exec_internal(cmd = 'docker', args = 'login', error = FALSE)
+  if (res[['status']] == 0) {
+    return(invisible(TRUE))
+  }
   psswrd_file <- tempfile()
   on.exit(file.remove(psswrd_file))
   msg <- paste0('Password for [', username, ']: ')
